Split lib/b/Prelude into Generation, Elimination and Transformation
