+++ benchmark/Streamly/Benchmark/Prelude/Serial.hs
@@ -30,7 +30,7 @@ main = do
               [ o_n_space_serial_toList size -- < 2MB
               , o_n_space_serial_outerProductStreams size
               , o_n_space_wSerial_outerProductStreams size
-              , o_n_space_serial_traversable size -- < 2MB
+              , o_n_space_serial size -- < 2MB
               , o_n_space_serial_foldr size
               ]
         ]
