+++ streamly.cabal
@@ -138,6 +138,7 @@ extra-source-files:
     benchmark/README.md
     benchmark/*.hs
     benchmark/lib/Streamly/Benchmark/*.hs
+    benchmark/lib/Streamly/Benchmark/Prelude/*.hs
     benchmark/Streamly/Benchmark/Memory/*.hs
     benchmark/Streamly/Benchmark/Data/*.hs
     benchmark/Streamly/Benchmark/Data/Prim/*.hs
